% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite.R
\name{generateFakeClusters}
\alias{generateFakeClusters}
\title{Generate fake cluster data}
\usage{
generateFakeClusters(nCluster = 2, nPoints = 100, nDim = 2,
  betweenSd = 3, withinSd = 1)
}
\arguments{
\item{nCluster}{number of clusters}

\item{nPoints}{number of points within each cluster}

\item{nDim}{number of dimensions}

\item{betweenSd}{cluster centers are normally distributed around 0 with this standard deviation}

\item{withinSd}{points are normally distributed around cluster centers with this standard deviation}
}
\value{
list with matrix giving point locations (rows are points, columns are dimensions), a matrix of cluster centers and vector of cluster IDs
}
\description{
Generate fake cluster data
}
\examples{
clusts<-generateFakeClusters()
plot(clusts$points,col=clusts$ids)
points(clusts$centers,pch='x',col=1:nrow(clusts$centers),cex=2)
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}

