% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite.R
\name{infiniteGaussian}
\alias{infiniteGaussian}
\title{Fit an infinite Bayesian mixture model to IPD and sequence data}
\usage{
infiniteGaussian(ipds, seqs, nIter = 100, maxK = 100, mc.cores = 4)
}
\arguments{
\item{ipds}{a matrix of IPD ratios with a row for each position of interest and a column for each surrounding base}

\item{seqs}{a matrix of single characters or vector of strings giving sequences of interest}

\item{nIter}{number of JAGS iterations}

\item{maxK}{the maximum number of clusters expected (smaller is less computationally expensive)}

\item{mc.cores}{number of cores (and chains) to use}
}
\value{
list with matrix giving simulated data and stored JAGS model (can be used for further iterations)
}
\description{
Fit an infinite Bayesian mixture model to IPD and sequence data
}
\examples{
fakeSeqs<-generateFakeClusters()
infiniteGaussian(fakeSeqs$ipds,fakeSeqs$seqs,maxK=10)
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}

