% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite.R
\name{seqToMat}
\alias{seqToMat}
\title{Convert sequences to a numbered matrix}
\usage{
seqToMat(seqs, bases = c("A", "C", "G", "T"))
}
\arguments{
\item{seqs}{a character vector of sequences}

\item{bases}{the potential characters in the sequences in the order to be numbered}
}
\value{
3-d logical array with the first dimension being sequence, second position and third base
}
\description{
Convert sequences to a numbered matrix
}
\examples{
seqToMat(c('AAAT','ATCG'))
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}

