% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite.R
\name{generateFakeClusters}
\alias{generateFakeClusters}
\title{Generate fake cluster data}
\usage{
generateFakeClusters(nCluster = 2, nPoints = 100, nIpd = 6, nBase = 3,
  nVariantBases = min(3, nBase), betweenSd = 3, withinSd = 1)
}
\arguments{
\item{nCluster}{number of clusters}

\item{nPoints}{number of points within each cluster}

\item{nIpd}{number of IPDs to generate}

\item{nBase}{number of bases to generate}

\item{nVariantBases}{number of bases in each PWM varying from equal proportions}

\item{betweenSd}{cluster centers are normally distributed around 0 with this standard deviation}

\item{withinSd}{points are normally distributed around cluster centers with this standard deviation}
}
\value{
list with matrix giving point locations (rows are points, columns are dimensions), a matrix of cluster centers, a vector of sequences, a list of PWMs (rows are positions, columns are bases) and vector of cluster IDs
}
\description{
Generate fake cluster data
}
\examples{
clusts<-generateFakeClusters()
plot(clusts$ipds,col=clusts$ids)
points(clusts$centers,pch='x',col=1:nrow(clusts$centers),cex=2)
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}

